% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2c.R
\name{p2c}
\alias{p2c}
\title{Convert Hole Positions to Coordinates}
\usage{
p2c(x, y)
}
\arguments{
\item{x}{an integer or a vector, the hole position(s).}

\item{y}{an integer, the number of rotor holes.}
}
\value{
\code{p2c} returns a vector of coordinates.
}
\description{
\code{p2c} returns the coordinates of rotor holes. This is not meant to be
called directly.
}
