% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie2.R
\name{pie2}
\alias{pie2}
\title{Draw a Rotor}
\usage{
pie2(
  x,
  labels = names(x),
  edges = 200,
  radius = 0.8,
  clockwise = FALSE,
  init.angle = if (clockwise) 90 else 0,
  density = NULL,
  angle = 45,
  col = NULL,
  border = NULL,
  lty = NULL,
  main = NULL,
  panel = NULL,
  ...
)
}
\arguments{
\item{x}{See \code{\link{pie}} for more details.}

\item{labels}{See \code{\link{pie}} for more details.}

\item{edges}{See \code{\link{pie}} for more details.}

\item{radius}{See \code{\link{pie}} for more details.}

\item{clockwise}{See \code{\link{pie}} for more details.}

\item{init.angle}{See \code{\link{pie}} for more details.}

\item{density}{See \code{\link{pie}} for more details.}

\item{angle}{See \code{\link{pie}} for more details.}

\item{col}{See \code{\link{pie}} for more details.}

\item{border}{See \code{\link{pie}} for more details.}

\item{lty}{See \code{\link{pie}} for more details.}

\item{main}{See \code{\link{pie}} for more details.}

\item{panel}{a letter for labeling panel.}

\item{...}{See \code{\link{pie}} for more details.}
}
\value{
\code{pie2} returns a figure.
}
\description{
\code{pie2} was modified after \code{\link{pie}}. This is not meant to be
called directly.
}
